deprecate svg prop, refactor to getSvg in utils.js
