+++ .dockerignore
@@ -7,3 +7,4 @@ npm-debug.log
 /icons.js
 /icons.mjs
 /icons.d.ts
+/utils.js
