+++ .gitignore
@@ -3,6 +3,7 @@
 /icons.js
 /icons.mjs
 /icons.d.ts
+/utils.js
 
 # Ignore all files in the icons folder
 icons/*
