+++ .npmignore
@@ -13,3 +13,6 @@
 !icons.js
 !icons.mjs
 !icons.d.ts
+!utils.js
+!utils.mjs
+!utils.d.ts
