+++ a.mjs
@@ -1,0 +1,5 @@
+import { getSvg } from './utils.mjs';
+import { siAwslambda } from './icons.mjs';
+const svg = getSvg(siAwslambda);
+
+console.log(svg);
